Pull in local static data for C file
