+++ config/splat.us.strtop.yaml
@@ -50,7 +50,7 @@ segments:
       - [0x670, .data, e_lock_camera]
       - [0x6B8, .data, e_breakable]
       - [0x7C8, .data, d_prize_drops]
-      - [0x80C, data]
+      - [0x80C, .data, 20B48]
       - [0x910, .data, st_update]
       - [0x95C, .data, collision]
       - [0xD1C, .data, e_red_door]
